apps/rie-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pandas==2.2.3
numpy==2.1.3
httpx==0.28.0
pydantic==2.10.3
pydantic-settings==2.6.1
aiosqlite==0.20.0
python-dotenv==1.0.1
scipy==1.14.1
